+++ docs/changelog.rst
@@ -1,6 +1,11 @@
 Change log
 =====================================
 
+1 May 2015
+-----------
+* 144: Max 100-150 words for description of resource - http://github.com/mpowering/django-orb/issues/issue/144
+
+
 30 Apr 2015
 -----------
 * 203: Add optional file size on ResourceURL object - http://github.com/mpowering/django-orb/issues/issue/203